requirements.txt
albumentations
gradio
numpy
pillow
pytorch_lightning
segmentation_models_pytorch
torchvision
